+++ src/plugins/footnotes/footnotes-utils.ts
@@ -1,0 +1,44 @@
+/*!
+ * © 2024 Atypon Systems LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {
+  InlineFootnoteNode,
+  ManuscriptNode,
+  schema,
+} from '@manuscripts/transform'
+import { ResolvedPos } from 'prosemirror-model'
+import { findChildren, findParentNodeClosestToPos } from 'prosemirror-utils'
+
+export const findTableInlineFootnoteIds = (
+  footnoteElement: ManuscriptNode,
+  $pos: ResolvedPos
+) => {
+  const tableElement = findParentNodeClosestToPos(
+    $pos,
+    (node) => node.type === schema.nodes.table_element
+  )?.node.firstChild
+
+  return tableElement
+    ? new Set(
+        findChildren(
+          tableElement,
+          (node) => node.type === schema.nodes.inline_footnote
+        )
+          .map(({ node }) => (node as InlineFootnoteNode).attrs.rids)
+          .flat()
+      )
+    : undefined
+}
